program that prints the first 50 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
  */
 int main(void)
 {
-int n1 = 1;
-int n2 = 2;
+long long  n1 = 1;
+long long n2 = 2;
 int i;
-printf("%d, %d", n1, n2);
+printf("%lld, %lld", n1, n2);
 for (i = 2; i < 50; i++)
 {
-int s = n1 + n2;
-printf(", %d", s);
+long long s = n1 + n2;
+printf(", %lld", s);
 n1 = n2;
 n2 = s;
 }
